Move the grid's specs to the grid component
The scroll-prefetch re-wiring and DOM-focus contracts now belong to
BrowseBinMemberGridComponent, so their tests move with them. What stays
on the shell is what the shell still owns: that an arrow key walks
previewId to the right member and hands that index to the grid — pinned
now for horizontal, vertical and clamped-at-the-end steps.
